price-monitor/backendAPI/requirements.txt
wheel
flask
flask-restful
google-cloud-firestore
gunicorn
